html/semantics/embedded-content/the-audio-element/audio-loading-lazy-window-onload.tentative.html

<!DOCTYPE html>
<link rel="author" title="Squarespace" href="https://www.squarespace.com/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<audio
  id="target"
  src="/media/sine440.mp3?pipe=trickle(d2)"
  loading="lazy"
  controls
></audio>

<script>
  async_test(t => {
    const audio = document.getElementById("target");

    window.addEventListener("load", t.step_func_done(() => {
      assert_less_than(
        audio.readyState,
        audio.HAVE_CURRENT_DATA,
        "Audio should not have finished loading before window.onload fired"
      );
    }));
  }, "Audio with loading=lazy does not delay window onload event");
</script>
