third_party/blink/web_tests/external/wpt/html/semantics/embedded-content/the-audio-element/audio-loading-lazy-window-onload.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Audio with loading=lazy does not delay window onload event
  assert_less_than: Audio should not have finished loading before window.onload fired expected a number less than 2 but got 4
Harness: the test ran to completion.
